Rename publisher-name-field file
