+++ src/components/title/_fields/publisher-name/index.js
@@ -1,1 +1,1 @@
-export { default } from './publisher-name';
+export { default } from './publisher-name-field';
